requirements.txt
flask>=2.2
pandas>=2.0
mysql-connector-python>=9.0
python-dotenv>=1.0
